l160:JWT Claims

+++ section11/code/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -29,7 +29,11 @@ class Item(Resource):
 
         return item.json(), 201  # code means that object has been created
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 404
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
